feat: Add the ability to display CasualOS videos in a <video> tag
